Change the type of the operand

+++ src/main/kotlin/visitors/number/RadonArithmeticMethodVisitor.kt
@@ -4,7 +4,7 @@ import org.objectweb.asm.MethodVisitor
 import org.objectweb.asm.Opcodes
 
 public class RadonArithmeticMethodVisitor(inner: MethodVisitor? = null) : MethodVisitor(Opcodes.ASM9, inner) {
-    private var operand: Long? = null
+    private var operand: Int? = null
 
     override fun visitInsn(opcode: Int) {
         when (opcode) {
